requirements.txt
transformers
torch
datasets
pandas
numpy
faiss-cpu
sentence-transformers
streamlit
accelerate
scikit-learn
